refactor testing decorators to work with pytest
